pyproject.toml
[tool.poetry]
name = "sumula"
version = "0.1.0"
description = ""
authors = ["Rodrigoneal <rodrigho2006@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^8.0.0"
tabula-py = {extras = ["jpype"], version = "^2.9.0"}
requests = "^2.31.0"
pypdf2 = "^3.0.1"
httpx = "^0.26.0"
pydantic = "^2.6.0"
fastapi = "^0.109.2"
uvicorn = "^0.27.0.post1"
bs4 = "^0.0.2"
python-dateutil = "^2.8.2"
sqlalchemy = "^2.0.27"
aiosqlite = "^0.19.0"
beanie = "^1.25.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
aiomysql = "^0.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.2.2"


[tool.poetry.group.ruff.dependencies]
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
